refactor: Landing page formatting changes
