+++ website/components/LandingPage/index.tsx
@@ -6,7 +6,6 @@ import { useNavigate } from "react-router";
 import { useSearchParams } from "react-router-dom";
 import styled from "styled-components";
 
-import { landingPageContent } from "./content";
 import Header from "../Header";
 import LoadModal from "../LoadModal";
 import { AppDataProps, DatasetEntry, ProjectEntry } from "../../types";
@@ -14,6 +13,8 @@ import { FlexColumnAlignCenter, FlexColumn, FlexRowAlignCenter, VisuallyHidden, 
 import { getArgsFromParams } from "../../utils/url_utils";
 import HelpDropdown from "../HelpDropdown";
 import { BannerVideo } from "../../assets/videos";
+
+import { landingPageContent } from "./content";
 
 const MAX_CONTENT_WIDTH_PX = 1060;
 
@@ -25,7 +26,7 @@ const Banner = styled(FlexColumnAlignCenter)`
   margin: 0;
 `;
 
-const BannerTextContainer = styled(FlexColumn)`
+const BannerTextContainer = styled(FlexColumnAlignCenter)`
   --padding-x: 30px;
   padding: 26px var(--padding-x);
   max-width: calc(${MAX_CONTENT_WIDTH_PX}px - 2 * var(--padding-x));
@@ -82,7 +83,7 @@ const FeatureHighlightsContainer = styled.li`
   grid-template-columns: repeat(auto-fit, minmax(230px, 1fr));
   padding: 0;
   justify-content: space-evenly;
-  gap: 12px 24px;
+  gap: 20px;
   margin: 30px 0 0 0;
 `;
 
@@ -123,7 +124,7 @@ const ProjectList = styled.ul`
     width: 100%;
     height: 1px;
     background-color: var(--color-layout-dividers);
-    margin-bottom: 10px;
+    margin-bottom: 15px;
   }
 `;
 
@@ -131,15 +132,21 @@ const ProjectCard = styled.li`
   display: flex;
   width: 100%;
   flex-direction: column;
-  gap: 12px;
+  gap: 8px;
 
   & h3 {
     font-weight: 600;
   }
 
   & p,
-  & h3 {
+  & h3,
+  & span {
     margin: 0;
+  }
+
+  & a {
+    // Add 2px margin to maintain the same visual gap that text has
+    margin: 2px 0 0 0;
   }
 `;
 
@@ -231,6 +238,7 @@ export default function LandingPage(): ReactElement {
   // TODO: Should the load buttons be link elements or buttons?
   // Currently both the link and the button inside can be tab-selected.
   const renderDataset = (dataset: DatasetEntry, index: number): ReactElement => {
+    // TODO: Use links here instead of button onClicks.
     return (
       <DatasetCard key={index}>
         <h4>{dataset.name}</h4>
@@ -288,10 +296,8 @@ export default function LandingPage(): ReactElement {
     return (
       <ProjectCard key={index}>
         {projectNameElement}
-        <FlexColumn $gap={6}>
-          <p>{project.description}</p>
-          {publicationElement}
-        </FlexColumn>
+        <p>{project.description}</p>
+        {publicationElement}
         {loadButton}
         {datasetList}
       </ProjectCard>
@@ -337,7 +343,7 @@ export default function LandingPage(): ReactElement {
         </BannerTextContainer>
       </Banner>
 
-      <ContentContainer $gap={10}>
+      <ContentContainer $gap={30} style={{ paddingBottom: "400px" }}>
         <FeatureHighlightsContainer>
           <FeatureHighlightsItem>
             <h3>Multiresolution OME-Zarr support</h3>
