+++ website/types.tsx
@@ -1,3 +1,5 @@
+import { ReactNode } from "react";
+
 import { AppProps } from "../src/aics-image-viewer/components/App/types";
 
 export type AppDataProps = Omit<AppProps, "appHeight" | "canvasMargin">;
@@ -10,7 +12,7 @@ export type DatasetEntry = {
 
 export type ProjectEntry = {
   name: string;
-  description: string;
+  description: ReactNode;
   publicationLink?: URL;
   publicationName?: string;
   loadParams?: AppDataProps;
